Safari-609.1.11.1/LayoutTests/http/tests/media/hls/hls-audio-tracks-locale-selection-expected.txt


RUN(internals.setUserPreferredLanguages(["jp", "fr", "es-EN", "en"]))
EVENT(canplaythrough)
EXPECTED (video.audioTracks.length == '3') OK
EXPECTED (video.audioTracks[0].enabled == 'false') OK
EXPECTED (video.audioTracks[1].enabled == 'true') OK
EXPECTED (video.audioTracks[2].enabled == 'false') OK
END OF TEST
